dkr/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
nltk==3.4.5
numpy==1.18.1
pandas==1.0.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
vader-sentiment==3.2.1.1
vaderSentiment==3.2.1
